u16 -> u32

+++ vibrato/src/common.rs
@@ -11,9 +11,9 @@ pub const fn bincode_config() -> config::Configuration<LittleEndian, Fixint> {
 
 /// The maximam length of an input sentence.
 ///
-/// Note that the value must be represented with u16 so that
-/// an (exclusive) end position can be represented in 16 bits.
+/// Note that the value must be represented with u32 so that
+/// an (exclusive) end position can be represented in 32 bits.
 pub const MAX_SENTENCE_LENGTH: usize = usize::MAX;
 
 /// The fixed connection id of BOS/EOS.
-pub const BOS_EOS_CONNECTION_ID: u16 = 0;
+pub const BOS_EOS_CONNECTION_ID: u32 = 0;
